Migrates the docs site to Nextra 4.6 / React 19 with live TSDoc-generated props
