+++ next.config.ts
@@ -10,10 +10,12 @@ const withNextra = nextra({
 
 const nextConfig: NextConfig = withNextra({
   reactStrictMode: true,
-  output: 'export',
   images: {
     unoptimized: true,
   },
+  // ts-morph (via reablocks-docs-theme/tsdoc) is Node-only; keep it out of the
+  // RSC/edge bundler so its optional deps don't trigger build warnings.
+  serverExternalPackages: ['ts-morph', '@ts-morph/common'],
 
   webpack(config) {
     // Grab the existing rule that handles SVG imports
@@ -39,6 +41,13 @@ const nextConfig: NextConfig = withNextra({
 
     // Modify the file loader rule to ignore *.svg, since we have it handled now.
     fileLoaderRule.exclude = /\.svg$/i;
+
+    // ts-morph's @ts-morph/common references `source-map-support` as an
+    // optional dep; webpack warns when it can't resolve it. Stub it out.
+    config.resolve.fallback = {
+      ...config.resolve.fallback,
+      'source-map-support': false,
+    };
 
     return config;
   },
