+++ package.json
@@ -6,11 +6,11 @@
   "scripts": {
     "dev": "next dev",
     "build": "next build",
+    "build:cloudflare": "next build && pagefind --site .next/server/app --output-path public/_pagefind && npx @cloudflare/next-on-pages",
+    "preview": "npx wrangler pages dev .vercel/output/static",
+    "postbuild": "pagefind --site .next/server/app --output-path public/_pagefind",
     "start": "next start",
-    "postinstall": "cp node_modules/reagraph/dist/docs/docgen.json src/data/reagraph-docgen.json",
-    "postbuild": "pagefind --site out --output-path out/_pagefind",
-    "lint": "next lint",
-    "preview": "npx wrangler pages dev out"
+    "lint": "next lint"
   },
   "browser": {
     "fs": false,
@@ -22,18 +22,17 @@
     "@mdx-js/react": "^3.1.0",
     "@types/react-syntax-highlighter": "^15.5.13",
     "motion": "^12.19.1",
-    "next": "^15.3.3",
-    "nextra": "^4.2.17",
-    "nextra-theme-docs": "^4.2.17",
+    "next": "^15.5.18",
+    "nextra": "^4.6.1",
     "postcss": "^8.5.6",
     "posthog-js": "^1.256.2",
     "posthog-node": "^5.2.1",
     "reablocks": "^9.1.1",
-    "reablocks-docs-theme": "^2.1.5",
+    "reablocks-docs-theme": "2.3.2",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-syntax-highlighter": "^15.6.1",
-    "reagraph": "^4.30.2",
+    "reagraph": "4.30.8",
     "tailwindcss": "^4.1.10"
   },
   "devDependencies": {
@@ -43,5 +42,11 @@
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
     "pagefind": "^1.3.0"
+  },
+  "pnpm": {
+    "overrides": {
+      "prismjs": "^1.30.0",
+      "postcss": "^8.5.10"
+    }
   }
 }
