--- src/app/docs/[[...mdxPath]]/page.tsx
@@ -1,25 +1,0 @@
-import { importPage, generateStaticParamsFor } from "nextra/pages";
-import { useMDXComponents } from "@/mdx-components";
-
-export const generateStaticParams = generateStaticParamsFor("mdxPath");
-
-export async function generateMetadata(props: any) {
-  const params = await props.params;
-  const { metadata } = await importPage(params.mdxPath);
-  metadata.title = `${metadata?.title} - Reagraph`;
-
-  return metadata;
-}
-
-const Wrapper = useMDXComponents().wrapper;
-
-export default async function Page(props: any) {
-  const params = await props.params;
-  const result = await importPage(params.mdxPath);
-  const { default: MDXContent, toc, metadata } = result;
-  return (
-    <Wrapper toc={toc} metadata={metadata}>
-      <MDXContent {...props} params={params} />
-    </Wrapper>
-  );
-}
