+++ src/content/skills.mdx
@@ -19,7 +19,17 @@ usage patterns.
 
 ## Installation
 
-Choose one of the following methods to make skills available in your project.
+The simplest way to install Reagraph skills is with the [`skills`](https://www.npmjs.com/package/skills) CLI — run it from your project root:
+
+```bash
+npx skills add reaviz/skills
+```
+
+This downloads the Reagraph skill tree into your project's `.claude/skills/` directory, where Claude Code will pick it up automatically.
+
+You can also browse and grab the skill from the [Skills Marketplace](https://skillsmp.com/skills/reaviz-skills-reagraph-skill-md).
+
+If you prefer to manage the files manually, choose one of the methods below instead.
 
 ### Option 1: Project-level (recommended for teams)
