--- src/data/reagraph-docgen.json.d.ts
@@ -1,13 +1,0 @@
-declare const metadata: Array<{
-  displayName: string;
-  props: {
-    [key: string]: {
-      description: string;
-      required: boolean;
-      type: { name: string };
-      defaultValue?: { value: string };
-    };
-  };
-}>;
-
-export default metadata;
