--- theme.config.tsx
@@ -1,13 +1,0 @@
-
-const config = {
-  logo: <span>Reagraph</span>,
-  project: {
-    link: 'https://github.com/reaviz/reagraph',
-  },
-  docsRepositoryBase: 'https://github.com/reaviz/reagraph',
-  footer: {
-    text: 'MIT 2024 © Reagraph.',
-  },
-};
-
-export default config;
