+++ wrangler.toml
@@ -1,6 +1,15 @@
 name = "reagraph-website"
 compatibility_date = "2024-09-23"
+compatibility_flags = ["nodejs_compat"]
 
-# Pages-specific configuration - point to static export output
-pages_build_output_dir = "out"
+# Cloudflare Pages output of `@cloudflare/next-on-pages` (run via `build:cloudflare`).
+pages_build_output_dir = ".vercel/output/static"
 
+[vars]
+NODE_ENV = "production"
+# Cloudflare applies the [vars] above to the BUILD environment too, and
+# `NODE_ENV=production` makes `pnpm install` skip devDependencies — which breaks
+# the build (no typescript/tailwind/pagefind/@svgr). This override forces pnpm to
+# install devDependencies anyway, so the build keeps working while the deployed
+# worker still gets NODE_ENV=production at runtime.
+NPM_CONFIG_PRODUCTION = "false"
